modify to long

+++ src/sample/Counter.java
@@ -2,5 +2,5 @@ package sample;
 
 public interface Counter {
 
-	public int count(int input, int checkNum);
+	public long count(int input, int checkNum);
 }
